algvex/requirements.txt
# ==============================================================================
# AlgVex Dependencies
# Version: 3.0.0
# Updated: 2025-01
# ==============================================================================

# ==================== Core ====================
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ==================== Quantitative Framework ====================
pyqlib==0.9.7            # Microsoft Qlib (2024.08 最新稳定版)
# hummingbot==2.11.0     # 单独安装，见 scripts/setup.sh (2025.12 最新版)

# ==================== Data Collection ====================
requests>=2.28.0
aiohttp>=3.8.0
ccxt>=4.4.0              # 交易所统一API
websockets>=11.0

# ==================== Machine Learning ====================
scikit-learn>=1.2.0
lightgbm>=4.0.0
xgboost>=1.7.0
joblib>=1.2.0

# ==================== Deep Learning (Optional) ====================
# torch>=2.0.0
# pytorch-lightning>=2.0.0

# ==================== Data Storage ====================
pyarrow>=12.0.0
fastparquet>=2023.0
sqlalchemy>=2.0.0
alembic>=1.10.0
asyncpg>=0.27.0
psycopg2-binary>=2.9.0

# ==================== API Framework ====================
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ==================== Task Queue ====================
celery>=5.3.0
redis>=4.5.0

# ==================== Configuration ====================
python-dotenv>=1.0.0
pyyaml>=6.0

# ==================== Logging ====================
loguru>=0.7.0

# ==================== Visualization ====================
matplotlib>=3.7.0
plotly>=5.14.0

# ==================== Testing ====================
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0

# ==================== Development ====================
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.2.0
pre-commit>=3.2.0

# ==================== Security ====================
cryptography>=40.0.0

# ==================== Monitoring ====================
prometheus-client>=0.16.0
